tempus-login-mailer/requirements.txt
customtkinter>=5.2.0
openpyxl>=3.1.0
html-for-docx>=1.1.0
python-docx>=1.1.0
mammoth>=1.8.0
pywin32>=306;sys_platform=="win32"
